[refactor] Main 함수 위치 수정 및 Option 엔티티 테이블 명 수정
