+++ src/main/java/com/example/SpotApplication.java
@@ -1,4 +1,4 @@
-package com.example.spot;
+package com.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
