+++ src/main/java/com/example/controller/TestController.java
@@ -12,5 +12,4 @@ public class TestController {
     public String test() {
         return "Hello World!";
     }
-
 }
