+++ src/main/java/com/example/domain/mapping/StudyPostComment.java
@@ -1,8 +1,7 @@
 package com.example.domain.mapping;
 
 import com.example.domain.common.BaseEntity;
-import com.example.domain.mapping.StudyLikedComment;
-import com.example.domain.mapping.StudyPost;
+import com.example.domain.study.StudyPost;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
